componentized case studies pages
