+++ src/index.tsx
@@ -4,8 +4,9 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import { Homepage } from './pages/Home'
 import './styles/main.css'
 import { About } from './pages/About'
-import { CaseStudies } from 'pages/CaseStudies'
-import { CaseStudyAlgorithm } from 'pages/CaseAlgorithm'
+import { CaseStudyAlgorithm } from 'pages/CaseStudies/clients/algorithm/CaseAlgorithm'
+import { CaseStudies } from 'pages/CaseStudies/main'
+import { CaseStudyPCT } from 'pages/CaseStudies/clients/pct/CasePCT'
 
 const container = document.getElementById('root') as HTMLDivElement
 const root = createRoot(container)
@@ -17,6 +18,7 @@ root.render(
       <Route path="/about" element={<About />}></Route>
       <Route path="/case-studies" element={<CaseStudies />}></Route>
       <Route path="/algorithm-agency" element={<CaseStudyAlgorithm />}></Route>
+      <Route path="/pct" element={<CaseStudyPCT />}></Route>
     </Routes>
   </BrowserRouter>
 )
