--- src/pages/CaseAlgorithm.tsx
@@ -1,148 +1,0 @@
-import { Dedicated } from 'components/Dedicated'
-import { Testimonials } from 'components/shared sections/Testimonials'
-import { Testimonial } from 'components/shared sections/Testimonials/Testimonial'
-import { BodyText } from 'components/typography/BodyText'
-import { Heading } from 'components/typography/Heading'
-import { Layout } from 'containers/layout/Layout'
-import { Request1, RequestAvatar1, Results1 } from 'public/images'
-export function CaseStudyAlgorithm() {
-  return (
-    <Layout footerVariant="delivers">
-      <section className="flex flex-col-reverse mb-20 sm:flex-row sm:items-center sm:mx-16 sm:justify-between lg:mx-32">
-        <div className="w-full sm:w-[41%] mx-4">
-          <Heading level="2" className="mb-2 sm:text-4xl lg:mb-4">
-            Request
-          </Heading>
-          <BodyText className="text-sm lg:text-lg">
-            Algorithm Agency came to More Seconds because they needed a
-            dedicated technical resource that was able to integrate into their
-            digital operation as seamlessly as their in-house marketing team
-            members.
-          </BodyText>
-        </div>
-        <div className="relative w-full sm:w-[40%] px-4 mb-10">
-          <img src={Request1} alt="request" className="rounded-lg" />
-          <img
-            src={RequestAvatar1}
-            alt="avatar"
-            className="absolute w-24 border-4 rounded-full -bottom-2 border-light left-2 sm:-left-8 sm:bottom-8 lg:scale-150 lg:bottom-24"
-          />
-        </div>
-      </section>
-      <section className="flex flex-col-reverse mb-20 sm:flex-row-reverse sm:justify-between sm:mx-16 lg:mx-32 lg:items-center">
-        <div className="w-full mx-4 sm:w-[50%] lg:w-[40%]">
-          <Heading level="2" className="mb-2 sm:text-4xl lg:mb-4">
-            Results
-          </Heading>
-          <div className="grid gap-4">
-            <BodyText className="text-sm leading-6 lg:text-lg">
-              ● Extended in-house development and technical operations knowledge
-              base
-            </BodyText>
-            <BodyText className="text-sm leading-6 lg:text-lg">
-              ● Provided a suite of on-demand web services including research
-              &#38; analysis, live web troubleshooting, web development
-              expertise, project management, and much more
-            </BodyText>
-            <BodyText className="text-sm leading-6 lg:text-lg">
-              ● Increased month-to-month ROI by expanding their premium client
-              web service offerings
-            </BodyText>
-            <BodyText className="text-sm leading-6 lg:text-lg">
-              ● Enabled marketing team to focus on their core competencies
-              without worrying about the technical details
-            </BodyText>
-          </div>
-        </div>
-        <div className="relative w-full sm:w-[40%] sm:h-full px-4 mb-10 sm:mb-0">
-          <img src={Results1} alt="request" className="h-full rounded-lg" />
-        </div>
-      </section>
-      <section>
-        <Heading level="2" className="mb-5 text-4xl text-center">
-          Bistro Boudin
-        </Heading>
-        <Heading level="3" className="mb-5 text-lg text-center uppercase">
-          <span>bistro boudin </span>
-          <span className="text-[#3939FA]">in partnership</span>
-          <span> with algorithm agency</span>
-        </Heading>
-        <BodyText className="mx-4 text-sm text-center sm:mx-16 lg:mx-32">
-          Bistro Boudin is a nationally recognized San Francisco Bakery,
-          Restaurant, and Catering Company. More Seconds was contracted by
-          Algorithm Agency to redesign the Bistro Boudin Flagship website,
-          develop a customized dedicated web server, and handle the ongoing
-          maintenance &#38; upgrade needs the client presented. This site uses
-          Bridge + Elementor to provide a clean and easy way to create and
-          update its frontend. It also includes customized PHP page templates,
-          custom javascript for animating and rendering, as well as a host of
-          custom features and functionality to improve their ADA compliance,
-          page load time, and user experience.
-        </BodyText>
-      </section>
-      <section>
-        <Heading level="2" className="mb-5 text-4xl text-center">
-          State Fair Entertainment
-        </Heading>
-        <Heading
-          level="3"
-          className="mx-2 mb-5 text-lg font-normal text-center uppercase"
-        >
-          <span>State Fair Entertainment </span>
-          <span className="text-accent">in partnership</span>
-          <span> with algorithm agency</span>
-        </Heading>
-        <BodyText className="mx-4 text-center sm:mx-16 lg:mx-32">
-          State Fair Entertainment, producer of the Official Summer Fair of
-          L.A., was created around the idea that a large-scale, high quality
-          entertainment venue can promote positivity, be eco &#38; family
-          friendly and be an asset to the community in which it is held. More
-          Seconds received an urgent request from Algorithm Agency to build a
-          custom website for their new client State Fair Entertainment. By
-          working with More Seconds for all of their development and web
-          consulting needs, Algorithm Agency managed to generate over $4M+ in
-          ticket &#38; accessory sales for State Fair Entertainment for their
-          inaugural event.
-        </BodyText>
-      </section>
-      <section className="mx-4 sm:mx-16 lg:mx-32">
-        <ul className="grid grid-cols-2 gap-3 grid-row-2 sm:grid-cols-4 sm:grid-rows-1">
-          <li className="flex flex-col items-center h-32 p-4 rounded-2xl bg-darkermid lg:h-56 lg:p-10">
-            <div className="flex items-center h-full">
-              <BodyText className="text-4xl font-bold sm:font-normal sm:text-5xl lg:text-7xl">
-                15
-              </BodyText>
-            </div>
-            <BodyText className="text-xl">Lorem Ipsum</BodyText>
-          </li>
-          <li className="flex flex-col items-center h-32 p-4 rounded-2xl bg-darkermid lg:h-56 lg:p-10">
-            <div className="flex items-center h-full">
-              <BodyText className="text-4xl font-bold sm:font-normal sm:text-5xl lg:text-7xl">
-                43%
-              </BodyText>
-            </div>
-            <BodyText className="text-xl">Lorem Ipsum</BodyText>
-          </li>
-          <li className="flex flex-col items-center h-32 p-4 rounded-2xl bg-darkermid lg:h-56 lg:p-10">
-            <div className="flex items-center h-full">
-              <BodyText className="text-4xl font-bold sm:font-normal sm:text-5xl lg:text-7xl">
-                20+
-              </BodyText>
-            </div>
-            <BodyText className="text-xl">Lorem Ipsum</BodyText>
-          </li>
-          <li className="flex flex-col items-center h-32 p-4 rounded-2xl bg-darkermid lg:h-56 lg:p-10">
-            <div className="flex items-center h-full">
-              <BodyText className="text-4xl font-bold sm:font-normal sm:text-5xl lg:text-7xl">
-                1230
-              </BodyText>
-            </div>
-            <BodyText className="text-xl">Lorem Ipsum</BodyText>
-          </li>
-        </ul>
-      </section>
-      <Testimonials title="Generate millions in revenue with  the right website" />
-      <Dedicated />
-    </Layout>
-  )
-}
