+++ src/pages/CaseStudies/clients/Request.tsx
@@ -1,0 +1,32 @@
+import { BodyText } from 'components/typography/BodyText'
+import { Heading } from 'components/typography/Heading'
+import { Request1, RequestAvatar1 } from 'public/images'
+import { AvatarHighlight } from 'public/svgs'
+
+type Props = {
+  children: string
+  image: string
+  avatar: string
+}
+
+export function Request(props: Props) {
+  return (
+    <section className="flex flex-col-reverse mb-20 sm:flex-row sm:items-center sm:mx-16 sm:justify-between lg:mx-32">
+      <div className="w-full sm:w-[41%] mx-4">
+        <Heading level="2" className="mb-2 sm:text-4xl lg:mb-4">
+          Request
+        </Heading>
+        <BodyText className="text-sm lg:text-lg">{props.children}</BodyText>
+      </div>
+      <div className="relative w-full sm:w-[40%] px-4 mb-10">
+        <img src={props.image} alt="request" className="rounded-lg" />
+        <img
+          src={RequestAvatar1}
+          alt="avatar"
+          className="absolute w-24 border-4 rounded-full -bottom-2 border-light left-2 sm:-left-8 sm:bottom-8 lg:scale-150 lg:bottom-24"
+        />
+        <AvatarHighlight className="absolute z-40 scale-50 -bottom-14 -left-4 sm:-left-14 sm:-bottom-6 lg:-left-20 lg:bottom-4" />
+      </div>
+    </section>
+  )
+}
