+++ src/pages/CaseStudies/main/Hero.tsx
