+++ src/pages/CaseStudies/main/index.tsx
