+++ src/public/images/index.tsx
@@ -17,6 +17,9 @@ import Brian from './brian.png'
 import Request1 from './request_image1.png'
 import RequestAvatar1 from './request_avatar1.png'
 import Results1 from './results1.png'
+import Request2 from './request_image2.png'
+import RequestAvatar2 from './request_avatar2.png'
+import Results2 from './results2.png'
 
 export {
   AlgorithmLogo,
@@ -37,5 +40,8 @@ export {
   Brian,
   Request1,
   RequestAvatar1,
-  Results1
+  Request2,
+  RequestAvatar2,
+  Results1,
+  Results2
 }
