+++ src/public/svgs/index.tsx
@@ -37,6 +37,7 @@ import { ReactComponent as Sputnik } from './Sputnik.svg'
 import { ReactComponent as SmallSaturn } from './small_saturn.svg'
 import { ReactComponent as Rocket } from './Rocket.svg'
 import { ReactComponent as Highlight } from './highlight.svg'
+import { ReactComponent as AvatarHighlight } from './avatar_highlight.svg'
 
 export {
   MSLogo,
@@ -77,5 +78,6 @@ export {
   Sputnik,
   SmallSaturn,
   Rocket,
-  Highlight
+  Highlight,
+  AvatarHighlight
 }
